+++ index.d.ts
@@ -1,7 +1,12 @@
 export {default as find} from './find';
 export {default as findNext} from './findNext';
 export {default as findPrev} from './findPrev';
-export {default as insert} from './insert';
+export {
+  default as insert,
+  insertIfNotExists,
+  insertOrReplaceIfExists,
+  insertOrThrowIfExists,
+} from './insert';
 export {default as iterate} from './iterate';
 export {default as maxValue} from './maxValue';
 export {default as minValue} from './minValue';