+++ index.test-d.ts
@@ -9,7 +9,11 @@ import * as wbt from './index';
 import find from './find';
 import findNext from './findNext';
 import findPrev from './findPrev';
-import insert from './insert';
+import insert, {
+  insertIfNotExists,
+  insertOrReplaceIfExists,
+  insertOrThrowIfExists,
+} from './insert';
 import iterate from './iterate';
 import maxValue from './maxValue';
 import minValue from './minValue';
@@ -29,6 +33,9 @@ declare function cmpNumberAndString(a: number, b: string): number;
 
 // Basic usage
 expectType<types.ImmutableTree<string>>(insert<string>(stringTree, '', cmpStrings, actions.NOOP));
+expectType<types.ImmutableTree<string>>(insertIfNotExists<string>(stringTree, '', cmpStrings));
+expectType<types.ImmutableTree<string>>(insertOrReplaceIfExists<string>(stringTree, '', cmpStrings));
+expectType<types.ImmutableTree<string>>(insertOrThrowIfExists<string>(stringTree, '', cmpStrings));
 expectType<Generator<string, undefined, undefined>>(iterate<string>(stringTree));
 expectType<Generator<string, undefined, undefined>>(reverseIterate<string>(stringTree));
 expectType<types.ImmutableTree<string> | null>(find<string>(stringTree, '', cmpStrings));
@@ -45,6 +52,9 @@ expectType<types.ImmutableTree<string> | null>(findPrev<string, number>(stringTr
 
 // Wrong tree type
 expectError<types.ImmutableTree<number>>(insert<number>(stringTree, '', cmpNumbers, actions.NOOP));
+expectError<types.ImmutableTree<number>>(insertIfNotExists<number>(stringTree, '', cmpStrings));
+expectError<types.ImmutableTree<number>>(insertOrReplaceIfExists<number>(stringTree, '', cmpStrings));
+expectError<types.ImmutableTree<number>>(insertOrThrowIfExists<number>(stringTree, '', cmpStrings));
 expectError<Generator<number, undefined, undefined>>(iterate<number>(stringTree));
 expectError<Generator<number, undefined, undefined>>(reverseIterate<number>(stringTree));
 expectError<types.ImmutableTree<number> | null>(find<number>(stringTree, 0, cmpNumbers));
@@ -70,6 +80,9 @@ expectType<typeof find>(wbt.find);
 expectType<typeof findNext>(wbt.findNext);
 expectType<typeof findPrev>(wbt.findPrev);
 expectType<typeof insert>(wbt.insert);
+expectType<typeof insertIfNotExists>(wbt.insertIfNotExists);
+expectType<typeof insertOrReplaceIfExists>(wbt.insertOrReplaceIfExists);
+expectType<typeof insertOrThrowIfExists>(wbt.insertOrThrowIfExists);
 expectType<typeof iterate>(wbt.iterate);
 expectType<typeof reverseIterate>(wbt.reverseIterate);
 expectType<typeof maxValue>(wbt.maxValue);